+++ docs/CHANGELOG.rst
@@ -18,6 +18,7 @@ Changed
 Fixed
 -----
 - Use clean file name for gene sets from differential expressions
+- Fix saving estimated counts output in ``tximport_summarize.R`` script
 
 
 ===================