+++ docs/CHANGELOG.rst
@@ -1,3 +1,4 @@
+
 ##########
 Change Log
 ##########
@@ -19,6 +20,7 @@ Fixed
 -----
 - Use clean file name for gene sets from differential expressions
 - Fix saving estimated counts output in ``tximport_summarize.R`` script
+- Add ``config.yaml`` to methylation_arrays Dockerfile
 
 
 ===================