+++ docs/CHANGELOG.rst
@@ -15,6 +15,7 @@ Changed
 -------
 - Replace Bedtools with Samtools for BAM to FASTQ file format conversion
   in ``bamtofastq-paired`` process
+- Bump docker image version in ``methylation-array-sesame`` process
 
 Fixed
 -----