+++ docs/CHANGELOG.rst
@@ -7,9 +7,9 @@ All notable changes to this project are documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 
-==========
-Unreleased
-==========
+===================
+38.3.0 - 2021-08-16
+===================
 
 Changed
 -------